requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
python-dotenv>=1.0
Werkzeug>=3.0
twilio>=9.0
